first commit
adding only essential files for running phase2_ntuplemaking code. This area should be kept nice and clean so that it's git clonable. The old phase2_ntuplemaker repo can be kept in its messy state as a backup place for all the myriad versions of this phase2_ntuplemaking code that exist.